wip: #59 pages now redirect if user doesn't have necessary role
